app/requirements.txt
streamlit
langgraph
langchain
langchain-core
python-dotenv
requests
pydantic
openai
google-genai
typing-extensions
ddgs
